Make migration pass on production where agents_users was already dropped manually

+++ db/migrate/20230425084138_remove_unused_tables.rb
@@ -3,8 +3,8 @@
 class RemoveUnusedTables < ActiveRecord::Migration[7.0]
   def change
     # rubocop:disable Rails/ReversibleMigration
-    drop_table :agents_users
-    drop_table :delayed_jobs
+    drop_table :agents_users if table_exists?(:agents_users)
+    drop_table :delayed_jobs if table_exists?(:delayed_jobs)
     # rubocop:enable Rails/ReversibleMigration
   end
 end
